sm_table inherits from ResidualsCompliantRecordSet
